+++ tests/VObject/Property/BinaryTest.php
@@ -11,7 +11,7 @@ class BinaryTest extends \PHPUnit_Framework_TestCase {
      */
     function testMimeDir() {
 
-        $vcard = new VObject\Component\VCard();
+        $vcard = new VObject\Component\VCard(['VERSION' => '3.0']);
         $vcard->add('PHOTO', ['a', 'b']);
 
     }